requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
typing_extensions==4.12.2
yarl==1.17.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.17.1
